# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calendar:
    get:
      operationId: listCalendarEvents
      tags: [calendar]
      summary: List calendar events for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createCalendarEvent
      tags: [calendar]
      summary: Create a calendar event (syncs to Google Calendar if configured)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalendarEventBody"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /calendar/{id}:
    delete:
      operationId: deleteCalendarEvent
      tags: [calendar]
      summary: Delete a calendar event
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── VIDEO PROJECTS ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateCalendarEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        attendeeEmails:
          type: array
          items:
            type: string
        dealId:
          type: integer
          nullable: true
        contactId:
          type: integer
          nullable: true
        meetingUrl:
          type: string
          nullable: true
          description: Custom meeting link (Zoom, Teams, Webex, Calendly, etc.)
      required: [title, startTime, endTime]
    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        dealId:
          type: integer
          nullable: true
        contactId:
          type: integer
          nullable: true
        googleEventId:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        attendeeEmails:
          type: array
          items:
            type: string
        meetLink:
          type: string
          nullable: true
        meetingUrl:
          type: string
          nullable: true
          description: Custom meeting link (Zoom, Teams, Webex, Calendly, etc.)
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, startTime, endTime, attendeeEmails, createdBy, createdAt]

    # ── ARTIST SCHEMAS ────────────────────────────────────────────────────
...